requirements.txt
# Data processing
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2

# Machine Learning (GPU support)
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Text processing
python-Levenshtein==0.23.0
nltk==3.8.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Demo app
streamlit==1.29.0

# Utils
tqdm==4.66.1
joblib==1.3.2
